Add files via upload
This project is about the classification of EEG data. The EEG data is about the comparison between Low Workload and Medium Workload. The project helps predict the classification of this EEG data, which consists of 62 channels, 512 data points, and 360 samples that are equally divided into 0 and 1, representing if the data is classified as either low or medium workload. The project focuses on implementing the logistic model to find out the accuracy of classification if the deep learning architecture approach is not used. Then the project focuses on the implementation of a deep learning approach to figure out how deep learning architecture can improve the performance of the classification using the same EEG data. In the deep learning approach, Convolution Neural Network architecture is used. 5-fold cross-validation is used in the same Deep Learning approach to enhance training accuracy. Finally, the entire data is used to judge the CNN architecture accuracy with respect to 5-fold cross-validation accuracy.